+++ src/routes/api/carts.ts
@@ -4,14 +4,12 @@ import { ErrorHandler, NoBodyError } from "../../utils/errors.ts";
 import { JWTKEY } from "../../utils/secrets.ts";
 import { Cart, DiscountCheck } from "../../utils/types.ts";
 
-import { stringifyJSON } from "../../utils/utils.ts";
+import { Delete, Get, stringifyJSON } from "../../utils/utils.ts";
 import {
   CartSchema,
   DiscountCheckSchema,
   UuidSchema,
 } from "../../utils/validator.ts";
-
-import { Delete, Get } from "../../dataAccessLayer/cacheOrDb.ts";
 
 export default class CartRoutes {
   #carts: Router;
@@ -128,7 +126,6 @@ export default class CartRoutes {
         });
         ctx.response.headers.set("content-type", "application/json");
       } catch (error) {
-        console.log(error);
         const data = ErrorHandler(error);
         ctx.response.status = data.code;
         ctx.response.headers.set("content-type", "application/json");
@@ -144,16 +141,16 @@ export default class CartRoutes {
           id: ctx.params.id,
         });
 
-        const cart = await Get<Cart>({
+        /*const cart = await Get<Cart>({
           id: `cart_${ctx.params.id}`,
           promise: this.#Container.CartService.Get({
             id: ctx.params.id,
           }),
-        });
-
-        /*const cart = await this.#Container.CartService.Get({
-          id: ctx.params.id,
         });*/
+
+        const cart = await this.#Container.CartService.Get({
+          id: ctx.params.id,
+        });
 
         const token = await jwt.default.sign({
           cartId: cart.id,
@@ -183,14 +180,9 @@ export default class CartRoutes {
 
         const data = await Get<Cart>({
           id: `cart_${ctx.params.id}`,
-          promise: this.#Container.CartService.Get({
-            id: ctx.params.id,
-          }),
+          promise: this.#Container.CartService.Get({ id: ctx.params.id }),
         });
 
-        /*const data = await this.#Container.CartService.Get({
-          id: ctx.params.id,
-        });*/
         ctx.response.body = stringifyJSON({
           carts: data,
         });
@@ -218,9 +210,6 @@ export default class CartRoutes {
           }),
         });
 
-        /*await this.#Container.CartService.Delete({
-          id: ctx.params.id,
-        });*/
         ctx.response.status = 201;
         ctx.response.headers.set("content-type", "application/json");
       } catch (error) {
