+++ src/utils/utils.ts
@@ -1,3 +1,5 @@
+import container from "../services/mod.ts";
+
 /**
  * Returns a string.
  * @param {Record<string, unknown} data
@@ -12,4 +14,67 @@ export function stringifyJSON(
     (_, v) => typeof v === "bigint" ? v.toString() : v,
     2,
   );
+}
+
+export async function Get<T>(params: {
+  id: string;
+  promise: Promise<T>;
+}): Promise<T> {
+  try {
+    let data;
+    try {
+      data = await Promise.any([
+        container.CacheService.get<T>(params.id),
+        params.promise,
+      ]);
+    } catch (_error) {
+      data = await params.promise;
+    }
+
+    await container.CacheService.set({
+      id: params.id,
+      data: stringifyJSON(data),
+      expire: (60 * 10),
+    });
+
+    return data;
+  } catch (error) {
+    throw new Error("Could not get any data", {
+      cause: error,
+    });
+  }
+}
+
+export async function Update<T>(params: {
+  id: string;
+  promise: Promise<T>;
+}): Promise<T> {
+  try {
+    const data = await params.promise;
+
+    await container.CacheService.set({
+      id: params.id,
+      data: stringifyJSON(data),
+      expire: (60 * 10),
+    });
+
+    return data;
+  } catch (_error) {
+    throw new Error("Could not get any data");
+  }
+}
+
+export async function Delete(params: {
+  id: string;
+  promise: Promise<void>;
+}): Promise<void> {
+  try {
+    await params.promise;
+
+    await container.CacheService.delete(params.id);
+  } catch (error) {
+    throw new Error("Could not get any data", {
+      cause: error,
+    });
+  }
 }
